change place to get latest submission to summary page, because submission not needed for online code editor -> it get's it on its own simpler than catching all edge cases in running exercise
